requirements.txt
flask
flask-login
flask-migrate
flask-sqlalchemy
flask-wtf
python-dotenv
requests
google-cloud-storage
google-cloud-secret-manager
mysqlclient
google-cloud-datastore
